got prompts working again
