+++ index.js
@@ -4,6 +4,7 @@ const fs = require('fs');
 const shell = require('./src/Shell.js');
 const util = require('util');
 const Github = require('./src/Github.js');
+const { prompt, MultiSelect } = require('enquirer');
 
 const debuglog = util.debuglog('ghpr');
 
@@ -11,25 +12,63 @@ async function main() {
     await validateInGitRepo();
     const githubClient = await createGithubClient();
 
-    const repoLabels = await githubClient.getLabels();
+    const labels = await githubClient.getLabels();
 
-    /*
+    const prompts = {
+        title: {
+            type: 'input',
+            name: 'title',
+            message: 'Pull Request Title: ',
+            validate: function(text) {
+                if (text.length < 0) {
+                    return 'Must include a title.';
+                }
+                return true;
+            },
+        },
+        base: {
+            type: 'input',
+            name: 'base',
+            default: 'master',
+            message: 'Base Branch: ',
+        },
+        head: async function(base) {
+            return {
+                type: 'input',
+                name: 'head',
+                default: await shell.exec('git rev-parse --abbrev-ref HEAD'),
+                message: 'Head Branch: ',
+                validate: function(text) {
+                    if (base === text) {
+                        return 'Base and Head branch must be different.';
+                    }
+
+                    return true;
+                },
+            };
+        },
+    };
     debuglog('Prompting user for title...');
-    const { title } = await inquirer.prompt(prompts.title);
+    const { title } = await prompt(prompts.title);
     debuglog('Title: ', title);
 
     debuglog('Prompting user for base...');
-    const { base } = await inquirer.prompt(prompts.base);
+    const { base } = await prompt(prompts.base);
     debuglog('Base: ', base);
 
     debuglog('Prompting user for head...');
-    const { head } = await inquirer.prompt(prompts.head(base));
+    const { head } = await prompt(await prompts.head(base));
     debuglog('Head: ', head);
 
     debuglog('Prompting user for labels...');
-    const { labels } = await inquirer.prompt(prompts.labels(repoLabels));
+    const { selectedLabels } = await new MultiSelect({
+        message: 'Select Labels: ',
+        name: 'labels',
+        choices: labels,
+    }).run();
     debuglog('Labels: ', labels);
 
+    /*
     debuglog('Beggining render of template...');
     const body = await renderBody();
     debuglog('Pull Request Body: ', body);
@@ -40,13 +79,13 @@ async function createGithubClient() {
     const token = await getAccessToken();
     const { owner, repo } = await getGithubRemoteInfo();
 
-    return new Github(token, owner, repo)
+    return new Github(token, owner, repo);
 }
 
 async function validateInGitRepo() {
     debuglog('Validating in a git directory...');
     try {
-        await shell.exec('git rev-parse --is-inside-work-tree') 
+        await shell.exec('git rev-parse --is-inside-work-tree');
     } catch (e) {
         console.error('This is not a Git project.');
         process.exit(1);
@@ -60,13 +99,15 @@ async function getAccessToken() {
     result = process.env.GITHUB_TOKEN || token;
 
     if (!result) {
-        console.error('Could not retrieve GitHub Token',
-                      'Export the token as an environment variable GITHUB_TOKEN=<token>',
-                      'Or set it in your git config: git config --global github.token <token>')
-        process.exit(1)
+        console.error(
+            'Could not retrieve GitHub Token',
+            'Export the token as an environment variable GITHUB_TOKEN=<token>',
+            'Or set it in your git config: git config --global github.token <token>'
+        );
+        process.exit(1);
     }
 
-    return result
+    return result;
 }
 
 async function getGithubRemoteInfo() {
